requirements.txt
requests
py-cord
python-dotenv